ARS/projects/ARS-backend/main.py
import os
from dotenv import load_dotenv
from fastapi import FastAPI, HTTPException
from pydantic import BaseModel
from typing import List, Dict, Any

# Placeholder for AI and IPFS imports
# from .ai_model import predict_reputation_score, fetch_transaction_data
# from .ipfs_utils import upload_to_ipfs, get_from_ipfs

from algosdk.v2client import indexer

load_dotenv()

app = FastAPI(
    title="Algorand Reputation System Backend",
    description="AI-powered reputation scoring and Soulbound NFT management API."
)

ALGOD_SERVER = os.getenv("ALGOD_SERVER")
INDEXER_SERVER = os.getenv("INDEXER_SERVER")

if not ALGOD_SERVER or not INDEXER_SERVER:
    raise ValueError("ALGOD_SERVER and INDEXER_SERVER environment variables must be set.")

indexer_client = indexer.IndexerClient(indexer_token="", algod_address=INDEXER_SERVER)

# Placeholder for Pydantic models
class ReputationPredictionRequest(BaseModel):
    algorand_address: str

class ReputationPredictionResponse(BaseModel):
    algorand_address: str
    reputation_score: float
    message: str = "Reputation score calculated successfully."

# Placeholder for API endpoints
@app.post("/ai/predict-reputation", response_model=ReputationPredictionResponse)
async def calculate_user_reputation_placeholder(request: ReputationPredictionRequest):
    raise HTTPException(status_code=501, detail="Reputation prediction not yet implemented.")

@app.get("/health")
async def health_check():
    return {"status": "ok", "message": "ARS Backend is running (placeholder)."}
