ARS/projects/ARS-backend/pyproject.toml
[tool.poetry]
name = "ars-backend"
version = "0.1.0"
description = "FastAPI backend for Algorand Reputation System (ARS)"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.30.1" }
python-dotenv = "^1.0.1"
algosdk = "^2.2.1"
requests = "^2.32.3"
scikit-learn = "^1.5.0"
pandas = "^2.2.2"
numpy = "^1.26.4"
joblib = "^1.4.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
